Move Result model from landing_page_checker app to directory
Also rename items.py to taxonomy.py to make it less generic
