+++ directory/models/__init__.py
@@ -1,3 +1,4 @@
-from directory.models.items import *  # noqa: F403,F401
+from directory.models.entry import *  # noqa: F403,F401
 from directory.models.pages import *  # noqa: F403,F401
 from directory.models.snippets import *  # noqa: F403,F401
+from directory.models.taxonomy import *  # noqa: F403,F401
